feat(webui): enable route-level manual code-splitting to optimize bundle size
