+++ packages/webui/routes/projects/$projectId/index.tsx
@@ -1,12 +1,8 @@
 import { client } from '@/ui/api/client'
-import FileSystemManager from '@/ui/components/file-system-manager'
 import { ProjectFolderSkeleton } from '@/ui/components/loading-skeletons'
-import { useTeamContextStore } from '@/ui/stores/team-context'
-import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
-import { useEffect } from 'react'
 
-const projectInfoQueryOptions = (projectId: string) => ({
+export const projectInfoQueryOptions = (projectId: string) => ({
   queryKey: ['projects', projectId],
   queryFn: async () => {
     const res = await client.api.projects[':projectId'].$get({
@@ -17,47 +13,7 @@ const projectInfoQueryOptions = (projectId: string) => ({
   },
 })
 
-function ProjectPage() {
-  const { projectId } = Route.useParams()
-  const { teamId, ensureTeamIdForProject } = useTeamContextStore()
-
-  useEffect(() => {
-    ensureTeamIdForProject(projectId)
-  }, [projectId, ensureTeamIdForProject])
-
-  const { data: projectInfo } = useSuspenseQuery(projectInfoQueryOptions(projectId))
-
-  const rootFolderId = projectInfo?.rootFolder
-
-  const { data: rootFolder } = useQuery({
-    queryKey: ['folders', rootFolderId!],
-    queryFn: async () => {
-      const res = await client.api.folders[':folderId'].$get({
-        param: { folderId: rootFolderId! },
-      })
-      if (!res.ok) throw new Error('failed to fetch folder')
-      return await res.json()
-    },
-    enabled: !!rootFolderId,
-  })
-
-  if (!teamId || !rootFolderId || !rootFolder) {
-    return <ProjectFolderSkeleton />
-  }
-
-  return (
-    <FileSystemManager
-      teamId={teamId}
-      projectId={projectId}
-      projectName={projectInfo.name ?? ''}
-      assetId={rootFolderId}
-      rootFolderId={rootFolderId}
-    />
-  )
-}
-
 export const Route = createFileRoute('/projects/$projectId/')({
-  component: ProjectPage,
   loader: ({ context: { queryClient }, params: { projectId } }) =>
     queryClient.ensureQueryData(projectInfoQueryOptions(projectId)),
   pendingComponent: () => <ProjectFolderSkeleton />,
