+++ packages/webui/routes/share/$shareId/files/$fileId.tsx
@@ -1,26 +1,6 @@
-import { PublicShareManager } from '@/ui/components/public-share-manager'
 import { createFileRoute } from '@tanstack/react-router'
 
-function PublicShareFilePage() {
-  const { shareId, fileId } = Route.useParams()
-  const { start, version, compare, cmpLeft, cmpRight, cmpActive } = Route.useSearch()
-
-  return (
-    <PublicShareManager
-      shareId={shareId}
-      initialFileId={fileId}
-      startTime={start}
-      versionId={version}
-      compare={compare}
-      compareLeftId={cmpLeft}
-      compareRightId={cmpRight}
-      compareActiveSide={cmpActive}
-    />
-  )
-}
-
 export const Route = createFileRoute('/share/$shareId/files/$fileId')({
-  component: PublicShareFilePage,
   validateSearch: (search: Record<string, unknown>) => {
     return {
       start: search.start ? Number(search.start) : undefined,
