+++ packages/webui/tsr.config.json
@@ -1,4 +1,5 @@
 {
   "routesDirectory": "routes",
-  "generatedRouteTree": "routeTree.gen.ts"
+  "generatedRouteTree": "routeTree.gen.ts",
+  "autoCodeSplitting": true
 }
